diff --git a/destroyall/Makefile b/destroyall/Makefile
updated copyright

updated to version 0.4

diff --git a/destroyall/destroyall.cpp b/destroyall/destroyall.cpp

added xattr, chmod and sudo (force)

diff --git a/findexts/Makefile b/findexts/Makefile

fixed reference to stringlib

diff --git a/findexts/findexts.cpp b/findexts/findexts.cpp

updated date

diff --git a/go-multiren/build b/go-multiren/build

removed -s for compatibility

added build and test system

diff --git a/go-multiren/multiren.go b/go-multiren/multiren.go

added permissions normalization
shortmonthname to digits
pad day with 0's

+func NormalizePermissions(filename string, force bool) {
+//ShortMonthNametoDigits : change short month to padded number
+func ShortMonthNametoDigits(m string) string {
+//PadDay : add a zero to days 1 through 9
+func PadDay(d string) string {

correctly parse ".tar.bz2"

only modtime if redate is enabled

only add date if adddate is enabled

diff --git a/norepeats/norepeats.cpp b/norepeats/norepeats.cpp

added option to look at only files 1GB and less

removed 'load from cache' option

added force (adds sudo) option

diff --git a/nospaces/nospaces.cpp b/nospaces/nospaces.cpp

added force option and permissions

added option to undo space-removal

diff --git a/stringsfltr/stringsfltr.cpp b/stringsfltr/stringsfltr.cpp

filter for strings utility

detect possible virus

+++ destroyall/destroyall.cpp
@@ -188,7 +188,7 @@ bool InFiltersList(cStringList filters, cString target) {
 }
 
 
-void lookfordups(cString wildcard,bool usedirs, bool destroy_all_links, bool exec_only) {
+void lookfordups(cString wildcard,bool usedirs, bool destroy_all_links, bool exec_only, bool force) {
 
 	cerr<<"wildcard="<<wildcard<<endl;
 
@@ -225,8 +225,10 @@ void lookfordups(cString wildcard,bool usedirs, bool destroy_all_links, bool exe
 			if(!exec_only || isExecutive(thelist.GetName(i),exec_only)) {
 				cout<<endl<<"# "<<thelist.GetName(i)<<" matches criteria for deletion"<<endl;
 				
-				
+				cout<<"sudo xattr -c "<<BashFriendly(thelist.GetName(i))<<endl;
+				cout<<"sudo chmod -N "<<BashFriendly(thelist.GetName(i))<<endl;
 			//	if(!InFiltersList(filters,thelist.GetName(i))) {
+				if(force) cout<<"sudo ";
 					cout<<"rm -rvf "<<BashFriendly(thelist.GetName(i))<<endl;
 			//	} else {
 			//		cout<<"# skipping "<<BashFriendly(thelist.GetName(i))<<" because of filter"<<endl;
@@ -247,7 +249,7 @@ void lookfordups(cString wildcard,bool usedirs, bool destroy_all_links, bool exe
 
 int main(int argc, char* argv[]) {
 
-	cerr<<"April 23rd, 2014"<<endl;
+	cerr<<"Sept 5th, 2020"<<endl;
 	cArgs arguments(argc,argv,"-");
 	
 	
@@ -282,12 +284,12 @@ int main(int argc, char* argv[]) {
 	
 	
 	if(!wildcard.Contains(';')) {
-		 if(arguments.IsSet("d") || arguments.IsSet("b")) lookfordups(wildcard,true,destroy_all_links,arguments.IsSet("e"));
-		 if(!arguments.IsSet("d") || arguments.IsSet("b")) lookfordups(wildcard,false,destroy_all_links,arguments.IsSet("e"));
+		 if(arguments.IsSet("d") || arguments.IsSet("b")) lookfordups(wildcard,true,destroy_all_links,arguments.IsSet("e"),arguments.IsSet("f"));
+		 if(!arguments.IsSet("d") || arguments.IsSet("b")) lookfordups(wildcard,false,destroy_all_links,arguments.IsSet("e"),arguments.IsSet("f"));
 		 
 	} else {
 		wildcardlist.FromString(wildcard,';');
-		for(int i=0; i<wildcardlist.Length(); i++) lookfordups(wildcardlist[i],usedirs,destroy_all_links,arguments.IsSet("e"));
+		for(int i=0; i<wildcardlist.Length(); i++) lookfordups(wildcardlist[i],usedirs,destroy_all_links,arguments.IsSet("e"),arguments.IsSet("f"));
 	}
